tests/api/requirements.txt
# Python dependencies for API testing
requests>=2.31.0
python-dotenv>=1.0.0

